Update glutin to 0.30.0
The glutin 0.30.0 update decouples glutin from winit which
provides us with basis for a multithreaded renderer. This
also improves robustness of our configuration picking,
context creation, and surface handling.

As an example we're now able to start on systems without a vsync,
we don't try to build lots of contexts to check if some config works,
and so on.

That also brings us possibility to handle context losses, but that's
a future work.

Fixes #1268.

+++ alacritty/src/config/window.rs
@@ -1,10 +1,10 @@
 use std::fmt::{self, Formatter};
 use std::os::raw::c_ulong;
 
-use glutin::window::Fullscreen;
 use log::{error, warn};
 use serde::de::{self, MapAccess, Visitor};
 use serde::{Deserialize, Deserializer, Serialize};
+use winit::window::Fullscreen;
 
 use alacritty_config_derive::{ConfigDeserialize, SerdeReplace};
 use alacritty_terminal::config::{Percentage, LOG_TARGET_CONFIG};
@@ -121,9 +121,9 @@ impl WindowConfig {
     }
 
     #[inline]
-    pub fn padding(&self, scale_factor: f64) -> (f32, f32) {
-        let padding_x = (f32::from(self.padding.x) * scale_factor as f32).floor();
-        let padding_y = (f32::from(self.padding.y) * scale_factor as f32).floor();
+    pub fn padding(&self, scale_factor: f32) -> (f32, f32) {
+        let padding_x = (f32::from(self.padding.x) * scale_factor).floor();
+        let padding_y = (f32::from(self.padding.y) * scale_factor).floor();
         (padding_x, padding_y)
     }
